backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
py-pde==0.35.0
numpy==1.24.3
websockets==12.0
python-multipart==0.0.6
